config: Require -g or --gui option to start interactive config.
OSCam's config.sh script is getting more and more features that
are not being noticed, because when the script is called without
parameters the config dialog is started. This changes the default
behaviour to show the available parameters when 'config.sh' is
called without any.

The supported and documented way of running GUI config is by running
'make config', this continues to work as expected.

git-svn-id: https://svn.streamboard.tv/oscam@7000 4b0bc96b-bc66-0410-9d44-ebda105a78c1