dvbapi: Sent data as one packet.
Some time ago I've sent a patch for generic dvbapi under linux (boxtype=pc).
It is working fine but I think there is need for one important enhancement.
Currently it is sending each CA_SET_PID and CA_SET_DESCR as two separate
packets which are: the request integer and the data. I think it will be
much better to sent it in one packet.

At first it is not elegant to sent it this way, besides the UDP packet
order is not the think that we should rely on.

About client implementation: as far as I know there is currently two
clients using the api: vdr-plugin-dvbapi and ostryck MuMuDVB fork
(maybe tvheadend will also use it). I contacted ostryck and he said
there will be no problem with this small api change. Obviously also
no problem on vdr-plugin-dvbapi side.

Patch by manio posted in ticket #2719.

git-svn-id: https://svn.streamboard.tv/oscam@7314 4b0bc96b-bc66-0410-9d44-ebda105a78c1