Add per user fakedelay option.
This option can be used to override global fakedelay option on
a per user basis. You can set it to higher or lower value or
when set to 0 fakedelay is disabled for the user. By default
this option is set to -1 which means "take the value from
global fakedelay option".

WebIf was not updated and can't be used to edit this option
currently. Patches are welcome (keep in mind that anticascading
can be disabled and then this option does not exist).

git-svn-id: https://svn.streamboard.tv/oscam@7778 4b0bc96b-bc66-0410-9d44-ebda105a78c1