smartreader
Changed mutex lock. By init fase the system_lock will be used like before.
Once the specific pthread for the reader self is made. Only that reader based pthread mutex will be locked and unlocked when needed.

Latency for v2 and tripple was far to low and could lead to short packet event. That's one of the causes which may lead to T1 resync.

The v1 1.04 and 1.05 firmware does have a error. Latency can't be changed there and is on 1 ms. Which most likely the cause why we need so called magic values on irdeto card detect.

General read timeout for V2 and Tripple increased by 9 %. Due to firmware error.

git-svn-id: https://svn.streamboard.tv/oscam@9919 4b0bc96b-bc66-0410-9d44-ebda105a78c1