log: Remove oscam-log-xtr.h
I fail to see the point of having a header file which defines two variables
and it is included in only one file oscam-log.c

Even if these variables are used in other places other than oscam-log.c (and
they are not) using additional header file is again the wrong thing to do.

There should be no functional changes after this patch.

git-svn-id: https://svn.streamboard.tv/oscam@10158 4b0bc96b-bc66-0410-9d44-ebda105a78c1