Clean SO_REUSEPORT ifdefs.
Most (if not all) of these calls to set socket option SO_REUSEPORT are
bogus IMHO, but cleaning this mess is left for another time. For now just
remove create a function and use it instead of sprinkling ifdefs everywhere.

I'm 97.8% sure that all of these calls can be removed...

For more info about SO_REUSEPORT see:
  http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=c617f398edd4db2b8567a28e899a88f8f574798d
  http://stackoverflow.com/questions/3261965/so-reuseport-on-linux

git-svn-id: https://svn.streamboard.tv/oscam@10231 4b0bc96b-bc66-0410-9d44-ebda105a78c1