oscam-log: Split cs_log_int().
Split cs_log_int() into separate functions for logging single lines
and logging with buffer dumping. This is made mainly to avoid having
unused parameters in a single function.

As far as the users of the logging API are concerned - the API is
unchanged.

bloatcheck reports 12077 bytes saved from x86_64 oscam binary and
14044 bytes saved on dm500 binary.

git-svn-id: https://svn.streamboard.tv/oscam@10243 4b0bc96b-bc66-0410-9d44-ebda105a78c1