Rewrite ecm header whitelist code.
Change the structure that tracked ecm header whitelist and rewrite
the related code. Now it is a lot simpler (68 lines less code) and
closer to ecm whitelist parser and generator.

Tests module was updated with support for ecm header white list.

There is a positive effect on binary size as well (-440 bytes).

{{{
 add/remove: 4/2 grow/shrink: 3/1 up/down: 1639/-2079 (-440)
 function                                     old     new   delta
 chk_ecm_hdr_whitelist                          -     699    +699
 mk_t_ecm_hdr_whitelist                         -     555    +555
 static.ecm_ratelimit_findspace              2499    2685    +186
 ecm_hdr_whitelist_add                          -     113    +113
 matching_reader                             1973    2008     +35
 clear_ecm_hdr_whitelist                        -      34     +34
 static.ecm_ratelimit_check                  2728    2745     +17
 static.free_reader_ecm_headerwhitelist        66       -     -66
 mk_t_ecmheaderwhitelist                      669       -    -669
 ecmheaderwhitelist_fn                       1473     129   -1344
}}}

A negative is that after the rewrite the ecm header whitelist structure
takes a bit more memory than before.

git-svn-id: https://svn.streamboard.tv/oscam@10484 4b0bc96b-bc66-0410-9d44-ebda105a78c1