!!!! Better stream type identification based on descriptors
OSCam now detects the type of each stream pid (video, audio, subtitle, etc) based, not only on the "stream_type" value in the PMT, but on the stream descriptors as well. Stream type detection is crucial when using the extended cw and this patch fixes many problems with PowerVu channels which had no working audio streams. Stream type detection is now done once and is the same whether the "simple" cw or the extended cw are used.

all based on Nautilus7 great JOB !!!!

git-svn-id: https://svn.streamboard.tv/oscam@11450 4b0bc96b-bc66-0410-9d44-ebda105a78c1